requirements.txt
# ============================================
# CORE DEEP LEARNING
# ============================================
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# ============================================
# DATA PROCESSING & MANIPULATION
# ============================================
pandas>=2.0.0
scipy>=1.11.0
h5py>=3.8.0

# ============================================
# MEDICAL IMAGING LIBRARIES
# ============================================
nibabel>=5.0.0           # NIfTI medical image format
pydicom>=2.3.0           # DICOM medical image format
SimpleITK>=2.2.0         # Medical image processing

# ============================================
# IMAGE PROCESSING
# ============================================
opencv-python>=4.7.0
scikit-image>=0.21.0
Pillow>=9.5.0
albumentations>=1.3.0    # Advanced data augmentation

# ============================================
# MACHINE LEARNING & EVALUATION
# ============================================
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0  # Handle class imbalance (SMOTE, etc.)

# ============================================
# EXPLAINABILITY & INTERPRETABILITY
# ============================================
shap>=0.42.0             # SHAP values for feature importance
grad-cam>=1.4.0          # Grad-CAM visual explanations
captum>=0.6.0            # PyTorch interpretability

# ============================================
# MODEL OPTIMIZATION & DEPLOYMENT
# ============================================
onnx>=1.14.0             # ONNX model export
onnxruntime>=1.15.0      # ONNX inference
# tensorrt>=8.6.0        # TensorRT (requires NVIDIA GPU, manual install)

# ============================================
# VISUALIZATION
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0           # Interactive plots

# ============================================
# EXPERIMENT TRACKING & LOGGING
# ============================================
tensorboard>=2.13.0
tqdm>=4.65.0
wandb>=0.15.0            # Optional: Weights & Biases experiment tracking

# ============================================
# CONFIGURATION MANAGEMENT
# ============================================
pyyaml>=6.0
hydra-core>=1.3.0        # Advanced configuration management
python-dotenv>=1.0.0     # Environment variables

# ============================================
# JUPYTER & DEVELOPMENT
# ============================================
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.5.0

# ============================================
# TESTING (Optional)
# ============================================
pytest>=7.3.0
pytest-cov>=4.1.0

# ============================================
# UTILITIES
# ============================================
requests>=2.31.0
kaggle>=1.5.16           # Kaggle API for dataset download
